% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootparticlefilter.R
\name{bootstrap_particle_filter}
\alias{bootstrap_particle_filter}
\title{bootstrap_particle_filter}
\usage{
bootstrap_particle_filter(observations, model, theta, algorithmic_parameters)
}
\description{
This function implements the bootstrap particle filter. It returns log-evidence estimates, particles, and normalized weights.
}

