% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_particle_filter.R
\name{conditional_particle_filter}
\alias{conditional_particle_filter}
\title{conditional_particle_filter}
\usage{
conditional_particle_filter(observations, model, theta, Nx, path = NULL)
}
\description{
This function implements the conditional particle filter. If no conditioning path is provided, it performs a regular bootstrap particle filter. It returns log-evidence estimates, particles, normalized weights, tree of paths, and one sample final path.
}

