% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_distributions.R
\name{dnorminvchisq}
\alias{dnorminvchisq}
\title{dnorminvchisq}
\usage{
dnorminvchisq(theta, mu0, kappa0, nu0, sigma02, log)
}
\description{
Density of a Normal-Inverse-Chi-square (cf. Bayesian data analysis equation 3.6)
}

