% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_distributions.R
\name{fast_dmvnorm}
\alias{fast_dmvnorm}
\title{fast_dmvnorm}
\usage{
fast_dmvnorm(x, mean, covariance)
}
\description{
Fast Normal log-density using Rcpp (observations row-wise, dimensions column-wise)
}
