% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dmvnorm_transpose.R
\name{fast_dmvnorm_transpose_cholesky}
\alias{fast_dmvnorm_transpose_cholesky}
\title{fast_dmvnorm_transpose_cholesky}
\usage{
fast_dmvnorm_transpose_cholesky(x, mean, cholcovariance)
}
\description{
fast_dmvnorm_transpose_cholesky
}

