% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_distributions.R
\name{fast_rmvnorm}
\alias{fast_rmvnorm}
\title{fast_rmvnorm}
\usage{
fast_rmvnorm(nparticles, mean, covariance)
}
\description{
Fast samples from Normal using Rcpp (observations row-wise, dimensions column-wise)
}

