% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_distributions.R
\name{fast_rmvnorm_transpose}
\alias{fast_rmvnorm_transpose}
\title{fast_rmvnorm_transpose}
\usage{
fast_rmvnorm_transpose(nparticles, mean, covariance)
}
\description{
Fast samples from  Normal using Rcpp (dimensions row-wise, observations column-wise)
}

