% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rmvnorm_transpose.R
\name{fast_rmvnorm_transpose_cholesky}
\alias{fast_rmvnorm_transpose_cholesky}
\title{fast_rmvnorm_transpose_cholesky}
\usage{
fast_rmvnorm_transpose_cholesky(nparticles, mean, cholcovariance)
}
\description{
fast_rmvnorm_transpose_cholesky
}
