% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_variance_reduction.R
\name{get_additional_particles_smc2}
\alias{get_additional_particles_smc2}
\title{get_additional_particles_smc2}
\usage{
get_additional_particles_smc2(thetas, normw, PFs, t, observations, model,
  logtargetdensities, algorithmic_parameters)
}
\description{
(SMC2 version) Generates additional particles given current ones, in order to reduce the variance
of the associated particle estimators. Nc is the desired total number of particles
to be used in the computation / estimation (so about Nc - Ntheta new particles are generated,
used in the estimation, and then discarded). The generated particles are equally weigthed.
Note that in the continuous observations case, the particles (theta,xt) target the posterior-filtering
distribution p(theta, xt | y_1, ..., y_t), whereas in the discrete case, they must target
the posterior-one-step-predictive distribution p(theta, xt | y_1, ..., y_(t-1)).
}

