% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_batch_initial_proposal.R
\name{get_batch_initial_proposal}
\alias{get_batch_initial_proposal}
\title{get_batch_initial_proposal}
\usage{
get_batch_initial_proposal(observations, model, algorithmic_parameters)
}
\description{
This function provides an initial proposal targetting P(theta|Ybatch) based on fitting a Normal to a PMMH sample from P(theta|Ybatch)
}
