% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_batch_logevidence.R
\name{get_batch_logevidence}
\alias{get_batch_logevidence}
\title{get_batch_logevidence}
\usage{
get_batch_logevidence(observations, model, proposal, algorithmic_parameters)
}
\description{
This function computes the logevidence of a batch of observation Y1:b via importance sampling
}

