% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_derivatives_log_density.R
\name{get_CV_RBF}
\alias{get_CV_RBF}
\title{get_CV_RBF}
\usage{
get_CV_RBF(ys, kfold, range_lambda, range_sigma2, order)
}
\description{
Finds the optimal lambda and sigma2 for p'(y) or p''(y) via cross-validation.
The function assumes length(ys) is a multiple of kfold.
}

