% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_derivatives_log_density.R
\name{get_derivative_RBF}
\alias{get_derivative_RBF}
\title{get_derivative_RBF}
\usage{
get_derivative_RBF(ys, lambda, sigma2, order, ynew = NULL)
}
\description{
Estimates p'(y) or p''(y) via RBF regression.
The function assumes length(ys) is a multiple of kfold.
}
