% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_derivatives_log_density.R
\name{get_derivative_RBFlocal}
\alias{get_derivative_RBFlocal}
\title{get_derivative_RBFlocal}
\usage{
get_derivative_RBFlocal(ys, ystar, sigma2star, order)
}
\description{
Estimates derivative of p(y) up to order 2 by fitting a constant via a local loss.
}

