% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_numderiv.R
\name{get_derivatives_from_genD}
\alias{get_derivatives_from_genD}
\title{get_derivatives_from_genD}
\usage{
get_derivatives_from_genD(genD_output, dimY)
}
\description{
This function extracts the jacobian and the diagonal of the hessians from the output of genD (cf. numDeriv)
}

