% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_derivatives_log_density.R
\name{get_estimate_RBF}
\alias{get_estimate_RBF}
\title{get_estimate_RBF}
\usage{
get_estimate_RBF(y, ys, theta, sigma2)
}
\description{
Evaluate the RBF model at y given the fitted parameters and the training set ys.
}
