% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_iid_exponential.R
\name{get_model_iid_exponential}
\alias{get_model_iid_exponential}
\title{get_model_iid_exponential}
\usage{
get_model_iid_exponential(a, b)
}
\description{
Univariate iid exponential observations
}

