% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_iid_gaussian.R
\name{get_model_iid_gaussian}
\alias{get_model_iid_gaussian}
\title{get_model_iid_gaussian}
\usage{
get_model_iid_gaussian(mu0, kappa0, nu0, sigma02)
}
\description{
Univariate iid normal observations with unknown mean and variance
}

