% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_iid_gaussian_unknown_mean.R
\name{get_model_iid_gaussian_unknown_mean}
\alias{get_model_iid_gaussian_unknown_mean}
\title{get_model_iid_gaussian_unknown_mean}
\usage{
get_model_iid_gaussian_unknown_mean(muprior, sigma2prior)
}
\description{
Univariate iid Gaussian observations with unknown mean (degenerate linear gaussian SSM)
}

