% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_iid_gaussian_unknown_variance.R
\name{get_model_iid_gaussian_unknown_variance}
\alias{get_model_iid_gaussian_unknown_variance}
\title{get_model_iid_gaussian_unknown_variance}
\usage{
get_model_iid_gaussian_unknown_variance()
}
\description{
Univariate iid Gaussian observations with unknown variance (degenerate linear gaussian SSM)
}

