% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_iid_lognormal.R
\name{get_model_iid_lognormal}
\alias{get_model_iid_lognormal}
\title{get_model_iid_lognormal}
\usage{
get_model_iid_lognormal(mu0, kappa0, nu0, sigma02)
}
\description{
Univariate iid lognormal observations
}
