% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_kangarooExponential.R
\name{get_model_kangarooExponential}
\alias{get_model_kangarooExponential}
\title{get_model_kangarooExponential}
\usage{
get_model_kangarooExponential(timesteps = data_kangaroo["time", ],
  range_sigma = 10, range_tau = 10, range_r = 10)
}
\description{
Population model for counts of kangaroos: model 2 in Knape and Valpine (2012).
This model has 3 parameters: \code{sigma}, \code{tau}, \code{r}.
The priors are independent uniforms
on (0,\code{range_sigma}) for \code{sigma},
on (0,\code{range_tau}) for \code{tau},
on (-\code{range_r},\code{range_r}) for \code{r}.
}
