% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_kangarooLogistic.R
\name{get_model_kangarooLogistic}
\alias{get_model_kangarooLogistic}
\title{get_model_kangarooLogistic}
\usage{
get_model_kangarooLogistic(timesteps = data_kangaroo["time", ],
  stepsize = 0.001, range_sigma = 10, range_tau = 10, range_r = 10,
  range_b = 10)
}
\description{
Population model for counts of kangaroos: model 1 in Knape and Valpine (2012).
This model has 4 parameters: \code{sigma}, \code{tau}, \code{r}, \code{b}.
The priors are independent uniforms
on (0,\code{range_sigma}) for \code{sigma},
on (0,\code{range_tau}) for \code{tau},
on (-\code{range_r},\code{range_r}) for \code{r},
on (0,\code{range_b}) for \code{b}.
}

