% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_kangarooRandomwalk.R
\name{get_model_kangarooRandomwalk}
\alias{get_model_kangarooRandomwalk}
\title{get_model_kangarooRandomwalk}
\usage{
get_model_kangarooRandomwalk(timesteps = data_kangaroo["time", ],
  range_sigma = 10, range_tau = 10)
}
\description{
Population model for counts of kangaroos: model 3 in Knape and Valpine (2012).
This model has 3 parameters: \code{sigma}, \code{tau}.
The priors are independent uniforms
on (0,\code{range_sigma}) for \code{sigma},
on (0,\code{range_tau}) for \code{tau}.
}

