% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_lineargaussian_discreteprior.R
\name{get_model_lineargaussian_discreteprior}
\alias{get_model_lineargaussian_discreteprior}
\title{get_model_lineargaussian_discreteprior}
\usage{
get_model_lineargaussian_discreteprior()
}
\description{
This implements a univariate linear Gaussian model with discrete prior
}

