% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_lineargaussian_iid.R
\name{get_model_lineargaussian_iid}
\alias{get_model_lineargaussian_iid}
\title{get_model_lineargaussian_iid}
\usage{
get_model_lineargaussian_iid()
}
\description{
This implements a degenerate univariate linear Gaussian model with iid states
}

