% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_lineargaussiansimpler_discreteprior.R
\name{get_model_lineargaussiansimpler_discreteprior}
\alias{get_model_lineargaussiansimpler_discreteprior}
\title{get_model_lineargaussiansimpler_discreteprior}
\usage{
get_model_lineargaussiansimpler_discreteprior()
}
\description{
Univariate linear Gaussian model with one parameter and discrete prior
}

