% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_simplerlineargaussian.R
\name{get_model_simplerlineargaussian}
\alias{get_model_simplerlineargaussian}
\title{get_model_simplerlineargaussian}
\usage{
get_model_simplerlineargaussian()
}
\description{
This implements the univariate linear Gaussian model
}
