% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_SVLevy_multifactor_noleverage.R
\name{get_model_SVLevy_multifactor_noleverage}
\alias{get_model_SVLevy_multifactor_noleverage}
\title{get_model_SVLevy_multifactor_noleverage}
\usage{
get_model_SVLevy_multifactor_noleverage(timesteps, mu0mu = 0,
  sigma02mu = 10, mu0beta = 0, sigma02beta = 10, r0xi = 1/5,
  r0w2 = 1/5, r0lambda1 = 1, r0lambda2_1 = 1/2, alpha0w = 1,
  beta0w = 1)
}
\description{
Levy-driven Stochastic volatility model with multi-factor, as described by equations
(14) and (15) in Chopin, Jacob, Papaspiliopoulos (2013), and discussed in more detail
in Barndorff-Nielsen and Shephard (2002).
The states are Xt = (v1t, z1t, v2t, z2t).
}
