% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_SVLevy_multifactor_withleverage.R
\name{get_model_SVLevy_multifactor_withleverage}
\alias{get_model_SVLevy_multifactor_withleverage}
\title{get_model_SVLevy_multifactor_noleverage}
\usage{
get_model_SVLevy_multifactor_withleverage(timesteps, mu0mu = 0,
  sigma02mu = 10, mu0beta = 0, sigma02beta = 10, r0xi = 1/5,
  r0w2 = 1/5, r0lambda1 = 1, r0lambda2_1 = 1/2, alpha0w = 1,
  beta0w = 1, mu0rho1 = 0, sigma02rho1 = 10, mu0rho2 = 0,
  sigma02rho2 = 10)
}
\description{
Levy-driven Stochastic volatility model with multi-factor, as described by equations
(14) and (15) in Chopin, Jacob, Papaspiliopoulos (2013), and discussed in more detail
in Barndorff-Nielsen and Shephard (2002).
The states are Xt = (v1t, z1t, v2t, z2t, sum_e1t, sum_e2t), where sum_e1t and sum_e2t denote the
sum of the error variables involved in the generation of the vt's.
}
