% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_SVLevy_singlefactor.R
\name{get_model_SVLevy_singlefactor}
\alias{get_model_SVLevy_singlefactor}
\title{get_model_SVLevy_singlefactor}
\usage{
get_model_SVLevy_singlefactor(timesteps, mu0mu = 0, sigma02mu = 10,
  mu0beta = 0, sigma02beta = 10, r0xi = 1/5, r0w2 = 1/5, r0lambda = 1)
}
\description{
Levy-driven Stochastic volatility model with single factor, as described by equations
(13) and (14) in Chopin, Jacob, Papaspiliopoulos (2013), and discussed in more detail
in Barndorff-Nielsen and Shephard (2002).
The states are Xt = (vt, zt).
}

