% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_proposals.R
\name{get_proposal_independent_normal}
\alias{get_proposal_independent_normal}
\title{get_proposal_independent_normal}
\usage{
get_proposal_independent_normal()
}
\description{
Independent Normal proposal using fitted mean and covariance matrix
}

