% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_proposals.R
\name{get_proposal_mixture}
\alias{get_proposal_mixture}
\title{get_proposal_mixture}
\usage{
get_proposal_mixture(nclust = 5, maxattempts = 5)
}
\description{
Independent proposal from a fitted mixture of Normals with \code{nclust} components (default is 5).
If the fit is unsuccessful, return independent Normal proposal (see \code{get_independent_normal_proposal}).
}
