% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hscore.R
\name{hscore}
\alias{hscore}
\title{hscore}
\usage{
hscore(observations, model, algorithmic_parameters)
}
\description{
This function computes successive prequential Hyvarinen score by running the SMC or SMC^2 algorithm. It also computes the successive log-evidence as a by-product.
}

