% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hscore_continuous.R
\name{hscore_continuous}
\alias{hscore_continuous}
\title{hscore_continuous}
\usage{
hscore_continuous(observations, model, algorithmic_parameters)
}
\description{
This function computes successive prequential Hyvarinen score for continuous observations by running the smc^2 algorithm. It also computes the successive log-evidence as a by-product.
}
