% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hscore_discrete.R
\name{hscore_discrete}
\alias{hscore_discrete}
\title{hscore_discrete}
\usage{
hscore_discrete(observations, model, algorithmic_parameters)
}
\description{
This function computes successive prequential Hyvarinen score for discrete observations by running the smc^2 algorithm. It also computes the successive log-evidence as a by-product.
}
