% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hscore_discrete_no_tempering.R
\name{hscore_discrete_no_tempering}
\alias{hscore_discrete_no_tempering}
\title{hscore_discrete_no_tempering}
\usage{
hscore_discrete_no_tempering(observations, model, algorithmic_parameters)
}
\description{
This function computes successive prequential Hyvarinen score for discrete observations by running the smc^2 algorithm. It also computes the successive log-evidence as a by-product.
}

