% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscore_resume.R
\name{hscore_resume}
\alias{hscore_resume}
\title{hscore_resume}
\usage{
hscore_resume(RDSsave = NULL, savefilename = NULL,
  next_observations = NULL, new_algorithmic_parameters = NULL)
}
\description{
This function resumes a previously interrupted SMC / SMC2 run to compute the hscore.
It is a wrapper of the four different cases: continuous or discrete / SMC or SMC2.
It also computes the successive log-evidence as a by-product.
}
