% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyvarinenSSM-package.R
\docType{package}
\name{HyvarinenSSM-package}
\alias{HyvarinenSSM-package}
\alias{HyvarinenSSM}
\title{HyvarinenSSM}
\description{
...
}
\details{
...
}
\author{
Stephane Shao, Pierre E. Jacob
}
\keyword{package}
