% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_Nx.R
\name{increase_Nx}
\alias{increase_Nx}
\title{increase_Nx}
\usage{
increase_Nx(observations, t, model, thetas, PFs, Ntheta)
}
\description{
This function runs a CPF in order to increase Nx (the number of particles). The default new number of particles is set to 2*Nx.
}

