% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_Nx_no_tempering.R
\name{increase_Nx_no_tempering}
\alias{increase_Nx_no_tempering}
\title{increase_Nx_no_tempering}
\usage{
increase_Nx_no_tempering(observations, t, model, thetas, xnormW, trees,
  algorithmic_parameters)
}
\description{
This function runs a CPF in order to increase Nx (the number of particles). The default new number of particles is set to 2*Nx.
}

