% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_kalmanfilter.R
\name{KF_assimilate_one}
\alias{KF_assimilate_one}
\title{KF_assimilate_one}
\usage{
KF_assimilate_one(Yt, t, phi, psi, sigmaV2, sigmaW2, initial_mean, initial_var,
  KF_current)
}
\description{
This function performs one step of Kalman filtering
}
