% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_kalmanfilter.R
\name{KF_logdpredictive}
\alias{KF_logdpredictive}
\title{KF_logdpredictive}
\usage{
KF_logdpredictive(Yt, t, KF_current)
}
\description{
This function returns the predictive log-density of the next observation
}
