% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_kangarooExponential.R
\name{get_model_kangarooExponential}
\alias{get_model_kangarooExponential}
\title{model.kangarooExponential}
\usage{
get_model_kangarooExponential(range_r = 10)
}
\description{
This implements the exponential model in Knape et al. (2012)
}

