% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resampling.R
\name{multinomial_resampling_n}
\alias{multinomial_resampling_n}
\title{Multinomial resampling}
\usage{
multinomial_resampling_n(normalized_weights, N)
}
\value{
A vector of ancestors
}
\description{
Runs a multinomial resampling algorithm, taking a normalized weights and a number of desired draws
and returning a vector of ancestors
}
