% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_distributions.R
\name{rnorminvchisq}
\alias{rnorminvchisq}
\title{rnorminvchisq}
\usage{
rnorminvchisq(N, mu0, kappa0, nu0, sigma02)
}
\description{
Samples from a Normal-Inverse-Chi-square (cf. Bayesian data analysis equation 3.6)
}

