% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_gamma.R
\name{search_gamma}
\alias{search_gamma}
\title{search_gamma}
\usage{
search_gamma(current_gamma, f, objective, maxsteps = 1000, tolerance = 0.01)
}
\description{
This function finds the optimal gamma for tempering given some objective function f (e.g. ESS)
}

