% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{simulateData}
\usage{
simulateData(model, theta, nobservations)
}
\description{
This function generates artificial data from a model. 'additional_parameters' is a list of n lists, where n is the number of observations that are generated (examples of additional parameters include step size for simulating SDEs)
}
