% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{smc}
\alias{smc}
\title{smc}
\usage{
smc(observations, model, algorithmic_parameters)
}
\description{
This function runs the SMC algorithm, using tempering.
It also computes the log-evidence and the prequential Hyvarinen score (optional).
}
