% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc2.R
\name{smc2}
\alias{smc2}
\title{smc2}
\usage{
smc2(observations, model, algorithmic_parameters)
}
\description{
This function runs the SMC2 algorithm, using adapive Nx and tempering.
It also computes the log-evidence and the prequential Hyvarinen score (optional).
It is a wrapper of the function \code{smc2_} with a time budgeting and partial save feature.
}
