% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc2_resume.R
\name{smc2_resume}
\alias{smc2_resume}
\title{smc2_resume}
\usage{
smc2_resume(RDSsave = NULL, savefilename = NULL, next_observations = NULL,
  new_algorithmic_parameters = NULL)
}
\description{
This function resumes a previously interrupted SMC2 run.
It is a wrapper of the function \code{smc2_resume_} with a time budgeting and partial save feature.
}
