% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_resume.R
\name{smc_resume}
\alias{smc_resume}
\title{smc_resume}
\usage{
smc_resume(RDSsave = NULL, savefilename = NULL, next_observations = NULL,
  new_algorithmic_parameters = NULL)
}
\description{
This function resumes a previously interrupted SMC run.
It is a wrapper of the function \code{smc_resume_} with a time budgeting and partial save feature.
}
