% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcKF_sampler.R
\name{smcKF_sampler}
\alias{smcKF_sampler}
\title{smcKF_sampler}
\usage{
smcKF_sampler(observations, model, algorithmic_parameters)
}
\description{
This function runs the SMC algorithm, using tempering.
}

