% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resampling.R
\name{ssp_resampling_n}
\alias{ssp_resampling_n}
\title{SSP resampling}
\usage{
ssp_resampling_n(normalized_weights, u_vector)
}
\value{
A vector of ancestors
}
\description{
Runs an SSP resampling algorithm, taking a vector of normalized weights (length N) and a vector of uniform variables (length N)
and returning a vector of ancestors (length N)
}
