% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resampling.R
\name{systematic_resampling_n}
\alias{systematic_resampling_n}
\title{Systematic resampling}
\usage{
systematic_resampling_n(normalized_weights, N, u)
}
\value{
A vector of ancestors
}
\description{
Runs a systematic resampling algorithm, taking a vector of normalized weights, a number of desired draws, and a uniform variable
and returning a vector of ancestors
}
