% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_trees.R
\name{tree_getattributes}
\alias{tree_getattributes}
\title{tree_getattributes}
\usage{
tree_getattributes(tree)
}
\description{
Get all the attributes (as a \code{list}) of an RCpp tree object. The attributes are respectively:
number of x-particles (\code{N}); upper bound on total size of tree (\code{M});
dimension of each x-particle (\code{dimx}); 'number of time "insert" has been called (\code{nsteps});
vector of ancestor indices (\code{a_star}); vector of offspring counts (\code{o_star});
vector of particles (\code{x_star}); vector of leaf indices (\code{l_star}).
}
