% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_trees.R
\name{tree_reconstruct}
\alias{tree_reconstruct}
\title{tree_reconstruct}
\usage{
tree_reconstruct(tree_attributes_list)
}
\description{
Reconstruct an RCpp tree object from a \code{list} of attributes. The attributes are respectively:
number of x-particles (\code{N}); upper bound on total size of tree (\code{M});
dimension of each x-particle (\code{dimx}); 'number of time "insert" has been called (\code{nsteps});
vector of ancestor indices (\code{a_star}); vector of offspring counts (\code{o_star});
vector of particles (\code{x_star}); vector of leaf indices (\code{l_star}).
}

