% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_trees.R
\name{trees_getattributes}
\alias{trees_getattributes}
\title{trees_getattributes}
\usage{
trees_getattributes(trees)
}
\description{
This is a wrapper that takes a \code{list} of trees as input
and outputs a \code{list} of tree attributes \code{list} via the function \code{tree_getattributes}
}
