% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_trees.R
\name{trees_reconstruct}
\alias{trees_reconstruct}
\title{trees_reconstruct}
\usage{
trees_reconstruct(trees_attributes_list)
}
\description{
This is a wrapper that takes a \code{list} of tree attributes \code{list}
and outputs a \code{list} of reconstructed trees via the function \code{tree_reconstruct}
}

