% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_distributions.R
\name{fast_dmvnorm_transpose}
\alias{fast_dmvnorm_transpose}
\title{fast_dmvnorm_transpose}
\usage{
fast_dmvnorm_transpose(x, mean, covariance)
}
\description{
Fast Normal log-density using Rcpp (dimensions row-wise, observations column-wise)
}

