% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_iid_lognormal.R
\name{get_model_iid_lognormal}
\alias{get_model_iid_lognormal}
\title{get_model_iid_lognormal}
\usage{
get_model_iid_lognormal(prior_mu0, prior_sigma02, prior_a, prior_b)
}
\description{
Univariate iid lognormal observations
}

