% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_kangarooLogistic.R, R/get_model_kangarooRandomwalk.R
\name{get_model_kangarooLogistic}
\alias{get_model_kangarooLogistic}
\alias{get_model_kangarooRandomwalk}
\title{get_model_kangarooRandomwalk}
\usage{
get_model_kangarooLogistic(timesteps = data_kangaroo["time", ],
  range_b = 10)

get_model_kangarooRandomwalk(timesteps = data_kangaroo["time", ])
}
\description{
This implements the logistic model in Knape et al. (2012)

This implements the random walk model in Knape et al. (2012)
}

