% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_kangarooRandomwalk.R
\name{get_model_kangarooRandomwalk}
\alias{get_model_kangarooRandomwalk}
\title{get_model_kangarooRandomwalk}
\usage{
get_model_kangarooRandomwalk()
}
\description{
This implements the random walk model in Knape et al. (2012)
}

