% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_lineargaussian.R
\name{get_model_lineargaussian}
\alias{get_model_lineargaussian}
\title{get_model_lineargaussian}
\usage{
get_model_lineargaussian()
}
\description{
This implements the univariate linear Gaussian model
}

