% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_lineargaussiansimpler.R
\name{get_model_simplerlineargaussian}
\alias{get_model_simplerlineargaussian}
\title{get_model_simplerlineargaussian}
\usage{
get_model_simplerlineargaussian()
}
\description{
Univariate linear Gaussian model with 2 unknown parameters
(\code{phi}, \code{sigmaW2}).
Latent states: X[t] = \code{phi}*X[t-1] + N(0,\code{sigmaW2}).
Observations: Y[t] = \code{psi}*X[t] + N(0,\code{sigmaV2}).
}
