% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hscore.R
\name{hscore}
\alias{hscore}
\title{hscore}
\usage{
hscore(observations, model, algorithmic_parameters)
}
\description{
This function computes successive prequential Hyvarinen score by running SMC or SMC2.
It is a wrapper of the four different cases: continuous or discrete / SMC or SMC2.
It also computes the successive log-evidence as a by-product.
}

