% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc2_sampler.R
\name{smc2_sampler}
\alias{smc2_sampler}
\title{smc2_sampler}
\usage{
smc2_sampler(observations, model, algorithmic_parameters)
}
\description{
This function computes successive prequential Hyvarinen score for continuous observations by running the smc^2 algorithm. It also computes the successive log-evidence as a by-product.
}

