% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling-systematic.R
\name{systematic_resampling_n}
\alias{systematic_resampling_n}
\title{Systematic resampling}
\usage{
systematic_resampling_n(normalized_weights, N, u)
}
\value{
A vector of ancestors
}
\description{
Runs a systematic resampling algorithm, taking a uniform variable, normalized weights and a number of desired draws
and returning a vector of ancestors
}
